+++ ReleaseNotes.md
@@ -1,3 +1,11 @@
+# Release 3.0.2
+
+# OpenSSL 3.0.2
+
+  - Upgrade to OpenSSL 3.0.2
+  - Fix some of the introduced deprecations
+  - Add support for CMSContentInfo Encryption
+
 # Release 1.0.1
 
 - SDK 15